refactor(platform): extract cartridge config into config/cartridge.rs (#2830)
Move cartridge catalog + metadata/image-cache path parsing and the
resolved_metadata_db_path/resolved_image_cache_path/resolved_favorites_path
helpers (as an impl FrontendConfig block) into config/cartridge.rs. mod.rs
apply_args/apply_config_value are now thin orchestrators. No behavior change.
